requirement.txt
pyserial==3.4
websocket-server==0.4
